chore: move uow modules into their own sub-package
