+++ src/service_layer/bootstrap.py
@@ -3,9 +3,9 @@ from pathlib import Path
 
 from src.config.config import load_config
 from src.service_layer.http_client import HTTPClient
-from src.service_layer.publishing_uow import PublishingUoW
 from src.service_layer.service import Service
-from src.service_layer.sqlalchemy_uow import SQLAlchemyUoW
+from src.service_layer.unit_of_work.publishing_uow import PublishingUoW
+from src.service_layer.unit_of_work.sqlalchemy_uow import SQLAlchemyUoW
 
 
 def setup_logging():
