+++ src/service_layer/default_handlers.py
@@ -2,7 +2,7 @@ from typing import Any, Awaitable, Callable, List, Type
 
 from src.domain import commands, events
 from src.service_layer import handlers
-from src.service_layer.publishing_uow import PublishingUoW
+from src.service_layer.unit_of_work.publishing_uow import PublishingUoW
 
 type EventHandlers = dict[
     Type[events.Event], List[Callable[[Any, PublishingUoW], Awaitable[None]]]
