+++ src/service_layer/queue/command_queue.py
@@ -1,8 +1,8 @@
 from typing import Any, Awaitable, Callable, List, Type
 
 from src.domain.commands import Command
-from src.service_layer.publishing_uow import PublishingUoW
 from src.service_layer.queue.queue import TaskQueue
+from src.service_layer.unit_of_work.publishing_uow import PublishingUoW
 
 CommandCallback = Callable[[Any, PublishingUoW], Awaitable[Any]]
