+++ tests/integration/service_layer/fakes.py
@@ -4,7 +4,7 @@ from src.adapters.repository import AbstractRepository
 from src.adapters.tracking_repository import TrackingRepository
 from src.domain.commands import Command
 from src.domain.events import Event
-from src.service_layer.uow import AbstractUoW
+from src.service_layer.unit_of_work.uow import AbstractUoW
 from tests.unit.service_layer.fakes import FakeRepository
