+++ tests/unit/core/fake_handlers.py
@@ -7,8 +7,8 @@ from src.service_layer.default_handlers import (
     CommandHandlers,
     EventHandlers,
 )
-from src.service_layer.publishing_uow import PublishingUoW
 from src.service_layer.types import Message
+from src.service_layer.unit_of_work.publishing_uow import PublishingUoW
 
 Calls = List[Tuple[Type[Message], Optional[Any]]]
