+++ tests/unit/service_layer/test_handlers.py
@@ -5,7 +5,7 @@ import pytest
 from src.core.types import UUID, Model
 from src.domain import commands
 from src.service_layer.handlers import handle_create_task
-from src.service_layer.publishing_uow import PublishingUoW
+from src.service_layer.unit_of_work.publishing_uow import PublishingUoW
 from tests.integration.service_layer.fakes import FakeUoW
